ai_service/requirements.txt
fastapi
uvicorn
requests
pydantic
music21
python-dotenv
pytest
pytest-asyncio
pandas
pydantic
pytest
httpx
magenta
librosa
